Implementation Local available disk space #23

+++ src/main/java/ch/psc/domain/storage/service/FileStorage.java
@@ -11,7 +11,7 @@ public interface FileStorage {
   
   public List<Future<File>> download(List<File> files);
   
-  public int getAvailableStorageSpace();
+  public double getAvailableStorageSpace();
   
   public Tree<File> getFileTree();
